Move read user endpoint to routes
