+++ routes/users/index.js
@@ -1,3 +1,4 @@
 module.exports = {
-  create: require('./create')
+  create: require('./create'),
+  read: require('./read')
 }
